+++ subsys/bluetooth/controller/ll_sw/nordic/hal/nrf5/radio/radio_nrf5.h
@@ -48,5 +48,12 @@
 #define HAL_RADIO_PDU_LEN_MAX (BIT(8) - 1)
 #endif
 
+#include <nrf_peripherals.h>
+
+#if defined(PPI_PRESENT)
 #include "radio_nrf5_ppi.h"
+#endif
+#if defined(DPPI_PRESENT)
+#include "radio_nrf5_dppi.h"
+#endif
 #include "radio_nrf5_txp.h"
